move batch_to_dict and dict_to_batch to synthesize
